Adding setup.py and pyproject.toml and adjust imports to package.
