+++ setup.cfg
@@ -1,0 +1,53 @@
+[metadata]
+description = Bi-Directional Equivariant Long-Range DNA Sequence Modeling
+name = caduceus
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/kuleshov-group/caduceus
+platforms=any
+authors = Yair Schiff
+classifiers =
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
+
+[options]
+python_requires = >=3.8
+zip_safe = False
+package_dir=
+    =src
+packages = find:
+install_requires=
+    biopython
+    causal-conv1d
+    datasets
+    einops
+    universal-pathlib
+    genomic-benchmarks
+    huggingface-hub
+    hydra-core
+    mamba-ssm
+    omegaconf
+    lightning[extra]
+    rich
+    scikit-learn
+    timm
+    torchmetrics
+    transformers
+    triton
+
+[options.extras_require]
+tests =
+    pytest
+dev =
+    %(tests)s
+    pre-commit
+
+[options.packages.find]
+exclude =
+    tests
+    tests.*
+where=src
