+++ src/caduceus/tasks/encoders.py
@@ -1,0 +1,88 @@
+from torch import nn
+
+import caduceus.models.nn.utils as U
+import caduceus.utils as utils
+
+
+class Encoder(nn.Module):
+    """Encoder abstraction
+
+    Accepts a tensor and optional kwargs. Other than the main tensor, all other arguments should be kwargs.
+    Returns a tensor and optional kwargs.
+    Encoders are combined via U.PassthroughSequential which passes these kwargs through in a pipeline. The resulting
+    kwargs are accumulated and passed into the model backbone.
+    """
+
+    def forward(self, x, **kwargs):
+        """
+        x: input tensor
+        *args: additional info from the dataset (e.g. sequence lengths)
+
+        Returns:
+        y: output tensor
+        *args: other arguments to pass into the model backbone
+        """
+        return x, {}
+
+
+# For every type of encoder/decoder, specify:
+# - constructor class
+# - list of attributes to grab from dataset
+# - list of attributes to grab from model
+
+registry = {
+    "stop": Encoder,
+    "id": nn.Identity,
+    "embedding": nn.Embedding,
+    "linear": nn.Linear,
+}
+
+dataset_attrs = {
+    "embedding": ["n_tokens"],
+    "linear": ["d_input"],  # TODO make this d_data?
+    "class": ["n_classes"],
+    "time": ["n_tokens_time"],
+    "onehot": ["n_tokens"],
+    "conv1d": ["d_input"],
+    "patch2d": ["d_input"],
+}
+
+model_attrs = {
+    "embedding": ["d_model"],
+    "linear": ["d_model"],
+    "position": ["d_model"],
+    "class": ["d_model"],
+    "time": ["d_model"],
+    "onehot": ["d_model"],
+    "conv1d": ["d_model"],
+    "patch2d": ["d_model"],
+    "timestamp_embedding": ["d_model"],
+    "layer": ["d_model"],
+}
+
+
+def _instantiate(encoder, dataset=None, model=None):
+    """Instantiate a single encoder"""
+    if encoder is None:
+        return None
+    if isinstance(encoder, str):
+        name = encoder
+    else:
+        name = encoder["_name_"]
+
+    # Extract dataset/model arguments from attribute names
+    dataset_args = utils.config.extract_attrs_from_obj(
+        dataset, *dataset_attrs.get(name, [])
+    )
+    model_args = utils.config.extract_attrs_from_obj(model, *model_attrs.get(name, []))
+
+    # Instantiate encoder
+    obj = utils.instantiate(registry, encoder, *dataset_args, *model_args)
+    return obj
+
+
+def instantiate(encoder, dataset=None, model=None):
+    encoder = utils.to_list(encoder)
+    return U.PassthroughSequential(
+        *[_instantiate(e, dataset=dataset, model=model) for e in encoder]
+    )
